// src/shared/api/generated/models/updateSupportWorkspaceRolloutDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSupportWorkspaceRolloutDto {
  enabled: boolean;
  hardOff: boolean;
  /**
   * @minLength 3
   * @maxLength 500
   */
  reason: string;
}